エレキ虫を倒したときの演出を実装。

+++ src/user/DashLightingParticle.h
@@ -26,7 +26,7 @@ private:
 	std::shared_ptr<Model>m_model;
 	Transform m_transform;
 	int m_flame;
-	int m_timer;
+	float m_timer;
 
 	std::array<std::shared_ptr<TextureBuffer>, 3> m_elecTextureBuffer;
 };
