+++ src/user/EnemyWave.h
@@ -69,7 +69,7 @@ private:
 
 	std::vector<ENEMY_INFO::SPAWN_INFO> m_enemys;	// 敵データ
 	int m_waveStartFrame;							// ウェーブ開始フレーム
-	int m_nowWaveFrame;								// このWaveが始まってからの経過時間。
+	float m_nowWaveFrame;								// このWaveが始まってからの経過時間。
 	bool m_isBounusStageFlag;
 	bool m_startWaveFlag;
 public:
